initial consolidation of ACS scripts
